requirements.txt
streamlit==1.31.0
protobuf==3.20.3
langchain==0.0.352
langchain-google-genai==0.0.6
sentence-transformers==2.6.1
PyPDF2==3.0.1
fpdf==1.7.2
google-generativeai==0.3.2
Pillow==10.2.0
pytesseract==0.3.10
opencv-python==4.9.0.80
numpy==1.26.3
python-dotenv==1.0.0
supabase==2.0.3
scikit-learn>=1.3.0
pydantic==1.10.13
langchain-core==0.1.10
transformers==4.36.2
tokenizers==0.13.3
huggingface-hub>=0.16.4,<1.0.0